requirements.txt
Django>=3.1.3,<3.2.0
djangorestframework>=3.12.2,<3.13.0
psycopg2>=2.8.6,<2.9.0
Pillow>=8.0.1,<8.1.0
PyYAML>=5.3.1,<5.4.0

flake8>=3.8.4,<3.9.0

gunicorn>=20.0.4,<20.1.0